services/risk/requirements.txt
asyncpg>=0.29.0
redis>=5.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
